Now handles error during code download
This is a legitimate use case when trying to get the code for files that
were deleted. Git will generate an error. This error will now be the
content of the file which should be ok as it will lead to 0 complexity.

Will add a separate column if we need to.

Also: minor refactoring in test_core.py.